% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sample.R
\docType{data}
\name{data_sample}
\alias{data_sample}
\title{Pakistan Intellectual Capital}
\format{
A data frame of 1142 rows and 13 columns
\describe{
\item{X}
\item{S.}
\item{Teacher.Name}
\item{Department}
\item{Province.University.Located}
\item{Designation}
\item{Terminal.Degree}
\item{Graduated.from}
\item{Country}
\item{Year}
\item{Area.of.Specialization.Research.Interests}
\item{Other.Information}
}
}
\source{
https://www.kaggle.com/alexisbcook/pakistan-intellectual-capital
}
\usage{
data_sample
}
\description{
A dataset containing list of professors from Pakistan.
}
\keyword{datasets}
