% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_similarity_id.R
\name{text_similarity_id}
\alias{text_similarity_id}
\title{Text Similarity Id}
\usage{
text_similarity_id(
  input_text,
  space = FALSE,
  ignore_case = TRUE,
  score = 80,
  eps = 0.15
)
}
\arguments{
\item{input_text}{A vector of text that need to be cleaned}

\item{space}{A logical indicating whether to include space in the calculation. Default TRUE}

\item{ignore_case}{A logical indicating whether to ignore case. Default TRUE}

\item{score}{A numeric value indicating the score accepted as similar. Default 0}

\item{eps}{A numeric value to measure if text is close to each other, the larger the value indicating that we accept the similarity loosely. Default 0.15}
}
\value{
A data frame contain of the original text and the unique ID assigned to the text
}
\description{
Generate unique ID for identical text
}
\examples{
text_similarity_id(c("South Korea","germany","southkorea","Germany"))

}
\author{
Dessy Amirudin
}
