% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_similarity_score.R
\name{text_similarity_score}
\alias{text_similarity_score}
\title{Text Similarity Score}
\usage{
text_similarity_score(
  input_text,
  target_text,
  space = TRUE,
  ignore_case = TRUE,
  score = 0
)
}
\arguments{
\item{input_text}{A vector of text that need to be cleaned}

\item{target_text}{A vector of text as the standard text format}

\item{space}{A logical indicating whether to include space in the calculation. Default TRUE}

\item{ignore_case}{A logical indicating whether to ignore case. Default TRUE}

\item{score}{A numeric value indicating the score accepted as similar. Default 0}
}
\value{
A dataframe contain the origital text, target text and the score
}
\description{
Generate similarity score between two text
}
\examples{
text_similarity_score(c("germany","southkorea"),"South Korea")

}
\author{
Dessy Amirudin
}
