% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarRText.R
\docType{package}
\name{similaRText}
\alias{similaRText}
\alias{similaRText-package}
\title{similaRText: Similar text score}
\description{
Easily generate custom data frame summaries
}
\author{
\strong{Maintainer}: Dessy Amirudin \email{amirudind@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
